refactor: (nit) collapse loader-utils into loaders
